// src/types.ts
export type InputEntityOptions = {
  entity?: string;
  name?: string;
  attribute?: string;
  statistic?: StatisticType;
  period?: StatisticPeriod | "auto" | AutoPeriodConfig;
  unit_of_measurement?: string;
  internal?: boolean;
  show_value?:
    | boolean
    | {
        right_margin: number;
      };
  offset?: TimeDurationStr;
  time_offset?: TimeDurationStr;
  extend_to_present?: boolean;
  filters?: FilterInput[];
  on_legend_click?: Function;
  on_legend_dblclick?: Function;
  on_click?: Function;
};

export type InputConfig = {
  type: "custom:plotly-graph";
  /**
   * The time to show on load.
   * It can be the number of hour (e.g 12),
   * a duration string, e.g 100ms, 10s, 30.5m, 2h, 7d, 2w, 1M, 1y,
   * or relative time, i.e:
   *  * current_minute
   *  * current_hour
   *  * current_day
   *  * current_week
   *  * current_month
   *  * current_quarter
   *  * current_year
   */
  hours_to_show?: number | TimeDurationStr | RelativeTimeStr;
  /** Either a number (seconds), or "auto" */
  refresh_interval?: number | "auto"; // in seconds
  color_scheme?: ColorSchemeNames | ColorSchemeArray | number;
  title?: string;
  offset?: TimeDurationStr;
  entities: (InputEntityOptions & Partial<Plotly.Data>)[];
  defaults?: {
    entity?: Partial<Plotly.Data>;
    xaxes?: Partial<Plotly.Layout["xaxis"]>;
    yaxes?: Partial<Plotly.Layout["yaxis"]>;
  };
  on_dblclick?: Function;
  layout?: Partial<Plotly.Layout>;
  config?: Partial<Plotly.Config>;
  ha_theme?: boolean;
  raw_plotly_config?: boolean;
  significant_changes_only?: boolean; // defaults to false
  minimal_response?: boolean; // defaults to true
  disable_pinch_to_zoom?: boolean; // defaults to false
  touch_hover?: boolean; // defaults to false
  autorange_after_scroll?: boolean; // defaults to false
  preset?: string | string[];
};

export type EntityConfig = EntityIdConfig & {
  unit_of_measurement?: string;
  internal: boolean;
  show_value:
    | boolean
    | {
        right_margin: number;
      };
  offset: number;
  extend_to_present: boolean;
  filters: FilterFn[];
  on_legend_click: Function;
  on_legend_dblclick: Function;
  on_click: Function;
} & Partial<Plotly.Data>;

export type Config = {
  title?: string;
  hours_to_show: number;
  refresh_interval: number | "auto"; // in seconds
  offset: number;
  entities: EntityConfig[];
  layout: Partial<Plotly.Layout>;
  config: Partial<Plotly.Config>;
  ha_theme: boolean;
  raw_plotly_config: boolean;
  significant_changes_only: boolean;
  minimal_response: boolean;
  disable_pinch_to_zoom: boolean;
  touch_hover: boolean;
  visible_range: [number, number];
  on_dblclick: Function;
  autorange_after_scroll: boolean;
};